refactor: move shared/providers to shared/container directory
