+++ src/shared/container/providers/MailProvider/dtos/ISendMailDTO.ts
@@ -1,4 +1,4 @@
-import IParseMailTemplateDTO from '@shared/providers/MailTemplateProvider/dtos/IParseMailTemplateDTO';
+import IParseMailTemplateDTO from '@shared/container/providers/MailTemplateProvider/dtos/IParseMailTemplateDTO';
 
 interface IMailContact {
   name: string;
