+++ src/shared/container/providers/MailProvider/index.ts
@@ -1,0 +1,9 @@
+import { container } from 'tsyringe';
+
+import EtherealMailProvider from './implementations/EtherealMailProvider';
+import SESMailProvider from './implementations/SESMailProvider';
+
+export default {
+  ethereal: container.resolve(EtherealMailProvider),
+  ses: container.resolve(SESMailProvider),
+};
