+++ src/shared/container/providers/MailProvider/models/IMailProvider.ts
