+++ src/shared/container/providers/MailTemplateProvider/dtos/IParseMailTemplateDTO.ts
