+++ src/shared/container/providers/MailTemplateProvider/models/IMailTemplateProvider.ts
