+++ src/shared/container/providers/StorageProvider/implementations/DiskStorageProvider.ts
@@ -2,7 +2,7 @@ import path from 'path';
 import fs from 'fs';
 import uploadConfig from '@config/upload';
 
-import IStorageProvider from '@shared/providers/StorageProvider/models/IStorageProvider';
+import IStorageProvider from '@shared/container/providers/StorageProvider/models/IStorageProvider';
 
 class DiskStorageProvider implements IStorageProvider {
   public async saveFile(file: string): Promise<string> {
